refactor: move native_converters to pkg
