+++ controllers/vpdeployment_controller_test.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"time"
 
-	"github.com/fintechstudios/ververica-platform-k8s-operator/api/native_converters"
 	"github.com/fintechstudios/ververica-platform-k8s-operator/api/v1beta2"
 	mocks "github.com/fintechstudios/ververica-platform-k8s-operator/mocks/vvp/appmanager"
 	"github.com/fintechstudios/ververica-platform-k8s-operator/pkg/annotations"
 	appmanagerapi "github.com/fintechstudios/ververica-platform-k8s-operator/pkg/vvp/appmanager-api"
+	"github.com/fintechstudios/ververica-platform-k8s-operator/pkg/vvp/native_converters"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
