+++ controllers/vpnamespace_controller_test.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"time"
 
-	"github.com/fintechstudios/ververica-platform-k8s-operator/api/native_converters"
 	"github.com/fintechstudios/ververica-platform-k8s-operator/api/v1beta1"
 	mocks "github.com/fintechstudios/ververica-platform-k8s-operator/mocks/vvp/platform"
+	"github.com/fintechstudios/ververica-platform-k8s-operator/pkg/vvp/native_converters"
 	platformapi "github.com/fintechstudios/ververica-platform-k8s-operator/pkg/vvp/platform-api"
 
 	. "github.com/onsi/ginkgo"
