+++ pkg/vvp/native_converters/deployment_metadata_test.go
