+++ pkg/vvp/native_converters/deployment_spec_test.go
