+++ pkg/vvp/native_converters/namespace.go
